+++ packages/use/index.ts
@@ -1,3 +1,4 @@
 export * from './src/watch/index';
 export * from './src/whenever/index';
 export * from './src/wheneverEffectScope/index';
+export * from './src/useCountdown/index';