+++ vitest.config.ts
@@ -15,6 +15,7 @@ export default defineConfig(({ mode }) => {
     },
     test: {
       globals: true,
+      environment: 'jsdom',
       coverage: {
         provider: 'v8',
       },